# [5.4] Adding MessageSelectorTest to Translation test suite and refactoring MessageSelector