Hello,

i've created a new MessageSelectorTest class with the objective to refactor the getPluralIndex method of the class MessageSelector in the Translation. The test cover all the methods of MessageSelector.

I've replaced the numerous cases by few strpos checks, to improve readability, testing and maintenance. I've also created a getUniquePluralIndex private function, used for countries that have a unique pattern of pluralization, so i don't have to use strpos for them, it runs faster then.

i've also done performance checks to verify that my changes dont overload the translation process :

| Locale | Execution Time After | Execution Time Before | Difference|
| :-------------: | :-------------: | :-------------: | :-------------: |
| `az_AZ`  | 9.05E-6  | 4.05E-6 | -5.0E-6  |
| `bn_BD`  | 4.05E-6 | 9.05E-6 | +5.0E-6  |
| `zu_ZA` | 5.0E-6 | 1.69E-5  | +1.19E-5  |
| `fi` | 4.05E-6  | 6.91E-6 | +2.86E-6  |
| `fil_PH` | 7.15E-6 | 1.28E-5 | +5.72E-6  |
| `wa_BE` | 5.96E-6 | 1.31E-5  | +7.15E-6  |
| `hr_HR`  | 8.1E-6  | 1.4E-5 | +5.96E-6  |
| `uk_UA`  | 8.1E-6 | 1.4E-5 | +5.96E-6  |
| `zs_CZ` | 8.1E-6 | 1.31E-5  | +5.0E-6  |
| `ga_IE` | 8.1E-6  | 1.40E-5 | +5.96E-6  |
| `ar_AE` | 9.05E-6 | 1.28E-5 | +5.72E-6  |

it seems to improve performance a little bit.
i've also executed all the tests on php 5.6 and it seems to work fine :+1: 

> OK (2716 tests, 6209 assertions)



Tell me what you think of my changes, if they have drawbacks etc...
I'm new to this repo, so perhaps i've made some mistakes.

thanks.